// src/app/@core/api/model/posteDto.ts
/**
 * EPR api
 * The list of EPR project endpoints
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ProjetDto } from '././projetDto';
import { ZoneDto } from '././zoneDto';
import { ConstatDto } from '././constatDto';


export interface PosteDto { 
    id: number;
    status: PosteDto.StatusEnum;
    date: string;
    label: string;
    email: string;
    projet: ProjetDto;
    zone: ZoneDto;
    constats: Array<ConstatDto>;
}
export namespace PosteDto {
    export type StatusEnum = 'in progress' | 'finished';
    export const StatusEnum = {
        InProgress: 'in progress' as StatusEnum,
        Finished: 'finished' as StatusEnum
    };
}